#11 refactoring difficulty rating
